Improve status messages to declare the target while waiting
The user may have multiple rsync-continuous processes running in parallel
and if every process just says

  Waiting for files to syncronize to $TARGET (Press CTRL+C to quit) ...

it will be hard to figure out which process is the correct if one of those
needs to be stopped. Change the status messages to include $TARGET when
scripts has completed the last sync and starts to wait for more modified
files.